+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## beta-1.2
++ Add channel option to the AMGN.runCommand method, and integrate user permissions to this method. This method will not run a command for a user who doesn't have permission to run the command.
+~ Made modlogs message when AMGN is closed down go to every modlogs channel
++ Add modlogs message to every guild when network is started
+
 ## beta-1.1.1
 Fix whitelist/blacklist breaking if not used